+++ src/main/java/model/AttendanceBook.java
@@ -1,10 +1,10 @@
 package model;
 
 import dto.DismissalCrewDto;
-import dto.DismissalCrewSortPolicy;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
+import model.policy.DismissalCrewSortPolicy;
 
 public class AttendanceBook {
