+++ src/main/java/model/AttendanceHistory.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import model.policy.CampusOperatingPolicy;
 
 public class AttendanceHistory {
