+++ src/main/java/model/policy/DismissalCrewSortPolicy.java
@@ -1,8 +1,9 @@
-package dto;
+package model.policy;
 
 import static model.AttendanceType.ABSENT;
 import static model.AttendanceType.LATE;
 
+import dto.DismissalCrewDto;
 import java.util.Comparator;
 
 public enum DismissalCrewSortPolicy {
