+++ src/test/java/model/AttendanceBookTest.java
@@ -11,6 +11,7 @@ import java.time.LocalTime;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
+import model.policy.EducationTimePolicy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
